+++ src/databricks/labs/ucx/source_code/python_linter.py
@@ -2,6 +2,10 @@ from __future__ import annotations
 
 import ast
 import logging
+from collections.abc import Iterable
+
+from databricks.labs.ucx.source_code.base import Linter, Advice, Advisory
+
 
 logger = logging.getLogger(__name__)
 
@@ -64,3 +68,40 @@ class ASTLinter:
         visitor = MatchingVisitor(node_type, match_nodes)
         visitor.visit(self._root)
         return visitor.matched_nodes
+
+
+class PythonLinter(Linter):
+
+    def lint(self, code: str) -> Iterable[Advice]:
+        linter = ASTLinter.parse(code)
+        nodes = linter.locate(ast.Call, [("run", ast.Attribute), ("notebook", ast.Attribute), ("dbutils", ast.Name)])
+        return [self._convert_dbutils_notebook_run_to_advice(node) for node in nodes]
+
+    @classmethod
+    def _convert_dbutils_notebook_run_to_advice(cls, node: ast.AST) -> Advisory:
+        assert isinstance(node, ast.Call)
+        path = cls.get_dbutils_notebook_run_path_arg(node)
+        if isinstance(path, ast.Constant):
+            return Advisory(
+                'notebook-auto-migrate',
+                "Call to 'dbutils.notebook.run' will be migrated automatically",
+                node.lineno,
+                node.col_offset,
+                node.end_lineno or 0,
+                node.end_col_offset or 0,
+            )
+        return Advisory(
+            'notebook-manual-migrate',
+            "Path for 'dbutils.notebook.run' is not a constant and requires adjusting the notebook path",
+            node.lineno,
+            node.col_offset,
+            node.end_lineno or 0,
+            node.end_col_offset or 0,
+        )
+
+    @staticmethod
+    def get_dbutils_notebook_run_path_arg(node: ast.Call):
+        if len(node.args) > 0:
+            return node.args[0]
+        arg = next(kw for kw in node.keywords if kw.arg == "path")
+        return arg.value if arg is not None else None
